reorganize-streambar-react
