+++ client/react/frontpage/components/StreamBar/ChatBox/ChatBox.js
@@ -3,6 +3,8 @@
 import React from 'react';
 import cookie from 'react-cookie';
 import { Grid, Col } from 'react-bootstrap';
+import Message from './Message';
+import MessageForm from './MessageForm';
 import './ChatBox.scss';
 
 const socket = io();
@@ -133,148 +135,6 @@ const ChatBox = React.createClass({
             />
           </Col>
         </Grid>
-      </span>
-    );
-  },
-});
-
-/**
-Message component displaying a message object
-
-@prop messageID: id of message represented
-@prop text: text of message
-@prop user: username associated with message
-@prop viewing_user: username associated with viewing user
-@prop date: timestamp of message
-* */
-const Message = React.createClass({
-  getInitialState() {
-    return {};
-  },
-  reportMessage() {
-    $.post(
-      ReportMessageURL,
-      {
-        id: this.props.messageID,
-      },
-      result => {
-        if (result.success) {
-          this.setState({ reported: true });
-        }
-      }
-    );
-  },
-  render() {
-    const date = new Date(this.props.date);
-    const Month = date.getMonth() + 1;
-    const Day = date.getDay() + 1;
-    const Hour = date.getHours();
-    const Minute = date.getMinutes();
-    const Second = date.getSeconds();
-    return (
-      <div className="message">
-        <div
-          className={
-            this.props.user != this.props.viewing_user ? (
-              'their-message'
-            ) : (
-              'my-message'
-            )
-          }>
-          {/* Message Body */}
-          <div className="message-body-tag">
-            {// if it's an image, display the image
-            this.props.text.split(' ').map(word => {
-              if (word.length > 4 && word.substring(0, 4) == 'http') {
-                return (
-                  <span>
-                    <br />
-                    <a href={word} target="_blank">
-                      {imageURL(word) ? <img src={word} /> : ''}
-                      <br />
-                      {word}
-                    </a>
-                    <br />
-                  </span>
-                );
-              }
-              // otherwise just display the text
-              return `${word} `;
-            })}{' '}
-            {/* Report Message */
-            this.props.user != this.props.viewing_user && (
-              <button className="report-message" onClick={this.reportMessage}>
-                {this.state.reported ? 'GOT IT' : 'REPORT'}
-              </button>
-            )}
-          </div>
-        </div>
-        <br />
-        <div
-          id="message-username-tag"
-          style={
-            this.props.user == this.props.viewing_user ? (
-              { float: 'right' }
-            ) : (
-              { float: 'left' }
-            )
-          }>
-          <span className="message-username-tag-username">
-            {this.props.user}
-          </span>
-          {` ${/* + Month + "/" + Day + " " */
-
-          Hour}:${Minute}:${Second}`}
-        </div>
-      </div>
-    );
-  },
-});
-
-/**
-Form to submit new message
-
-@prop user: username associated with viewing user
-@prop onMessageSubmit: callback to execute with new message data
-* */
-const MessageForm = React.createClass({
-  getInitialState() {
-    return { text: '' };
-  },
-  handleSubmit(e) {
-    e.preventDefault();
-    if (this.state.text.trim() != '') {
-      const message = {
-        user: this.props.user,
-        text: this.state.text,
-        date: new Date(),
-      };
-      this.props.onMessageSubmit(message);
-      this.setState({ text: '' });
-    }
-  },
-  changeHandler(e) {
-    this.setState({ text: e.target.value });
-  },
-  render() {
-    return (
-      <span>
-        <div className="message_form">
-          <form onSubmit={this.handleSubmit}>
-            <input
-              onChange={this.changeHandler}
-              value={this.state.text}
-              placeholder="Write a Message"
-            />
-            <br />
-            <br />
-            <div id="load_more_wrapper">
-              <center>
-                <input type="submit" value="SEND" id="load_more" />
-              </center>
-            </div>
-          </form>
-        </div>
       </span>
     );
   },
