+++ client/react/frontpage/components/StreamBar/ChatBox/ChatBox.scss
@@ -1,10 +1,9 @@
 /**
 ChatBox.scss
-
 Sass styling for frontpage chat box
 **/
 
-@import '../common';
+@import '../../../common';
 
 .chat-box-fade-bottom {
   -webkit-mask-image: -webkit-gradient(linear, left 90%, left bottom, from(rgba(0,0,0,1)), to(rgba(0,0,0,0)));
