+++ client/react/frontpage/components/StreamBar/StreamBar.scss
